Refactored Recipe drop down and prepared column for showing percentage of recipe
